fixed granularity, still need to fix dd mode

+++ include/wsa_sweep_device.h
@@ -71,7 +71,7 @@ struct wsa_power_spectrum_config {
 	uint64_t fstop;
 
 	/// the rbw
-	double rbw;
+	uint64_t rbw;
 			
 	/// a sweep plan that achieves capturing the spectrum requested
 	struct wsa_sweep_plan *sweep_plan;
